backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gitpython==3.1.40
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.18
requests==2.31.0
python-dotenv==1.0.0
tiktoken==0.5.1
langchain==0.0.340
numpy==1.24.3
PyPDF2==3.0.1